requirements.txt
discord.py
discord.py-stubs
motor
pyYAML
jishaku
pillow
aiogtts
emojis
jamo
wand
